Clean up ocr2 helper functions
