+++ integration-tests/actions/keeper_benchmark_helpers.go
@@ -72,7 +72,7 @@ func DeployBenchmarkKeeperContracts(
 	registrar := DeployKeeperRegistrar(t, registryVersion, linkToken, registrarSettings, contractDeployer, client, registry)
 	if registryVersion == ethereum.RegistryVersion_2_0 {
 		nodesWithoutBootstrap := chainlinkNodes[1:]
-		ocrConfig := BuildKeepersOCR2Config(t, nodesWithoutBootstrap, *registrySettings, registrar.Address(), 5*blockTime)
+		ocrConfig := BuildAutomationOCR2Config(t, nodesWithoutBootstrap, *registrySettings, registrar.Address(), 5*blockTime)
 		err = registry.SetConfig(*registrySettings, ocrConfig)
 		require.NoError(t, err, "Registry config should be be set successfully")
 	}
