+++ integration-tests/chaos/automation_chaos_test.go
@@ -228,7 +228,7 @@ func TestAutomationChaos(t *testing.T) {
 
 			actions.CreateOCRKeeperJobs(t, chainlinkNodes, registry.Address(), network.ChainID, 0)
 			nodesWithoutBootstrap := chainlinkNodes[1:]
-			ocrConfig := actions.BuildKeepersOCR2Config(t, nodesWithoutBootstrap, defaultOCRRegistryConfig, registrar.Address(), 5*time.Second)
+			ocrConfig := actions.BuildAutomationOCR2Config(t, nodesWithoutBootstrap, defaultOCRRegistryConfig, registrar.Address(), 5*time.Second)
 			err = registry.SetConfig(defaultOCRRegistryConfig, ocrConfig)
 			require.NoError(t, err, "Registry config should be be set successfully")
 			require.NoError(t, chainClient.WaitForEvents(), "Waiting for config to be set")
