+++ integration-tests/smoke/automation_test.go
@@ -674,7 +674,7 @@ func TestAutomatedCheckPerformGasLimit(t *testing.T) {
 	// Now increase checkGasLimit on registry
 	highCheckGasLimit := automationDefaultRegistryConfig
 	highCheckGasLimit.CheckGasLimit = uint32(5000000)
-	ocrConfig := actions.BuildKeepersOCR2Config(t, nodesWithoutBootstrap, highCheckGasLimit, registrar.Address(), 5*time.Second)
+	ocrConfig := actions.BuildAutomationOCR2Config(t, nodesWithoutBootstrap, highCheckGasLimit, registrar.Address(), 5*time.Second)
 	err = registry.SetConfig(highCheckGasLimit, ocrConfig)
 	require.NoError(t, err, "Registry config should be be set successfully")
 	err = chainClient.WaitForEvents()
@@ -835,7 +835,7 @@ func setupAutomationTest(
 
 		actions.CreateOCRKeeperJobs(t, chainlinkNodes, registry.Address(), network.ChainID, 0)
 		nodesWithoutBootstrap := chainlinkNodes[1:]
-		ocrConfig := actions.BuildKeepersOCR2Config(t, nodesWithoutBootstrap, registryConfig, registrar.Address(), 5*time.Second)
+		ocrConfig := actions.BuildAutomationOCR2Config(t, nodesWithoutBootstrap, registryConfig, registrar.Address(), 5*time.Second)
 		err = registry.SetConfig(automationDefaultRegistryConfig, ocrConfig)
 		require.NoError(t, err, "Registry config should be be set successfully")
 		require.NoError(t, chainClient.WaitForEvents(), "Waiting for config to be set")
